Move DataSumFinder to its final location. Add reference to the CommandLineParser
